Add snippets for timing out a job.
Added a line to the end of the example job scripts because sge will not give an error when a job times out. This way you can obsere whether the job finished successfully or not. Another way would be to check the exit code from `jobhist`, but that is more advanced.